refactor: organized types into seperate module
