+++ backend/src/models/types/index.ts
@@ -1,2 +1,16 @@
 export type AutoSetFields = "createdOn" | "lastModifiedOn" | "ttl";
 export type ApiResponseUrlType = "GS_PATH" | "API_URI";
+
+export interface MultiSizePhoto {
+  small: string;
+  medium: string;
+  large: string;
+}
+
+export type Language = "ENGLISH" | "BANGLA" | "HINDI";
+
+export type IdentityType = "OWNER" | "TENANT";
+
+export type AcceptGender = "MALE" | "FEMALE" | "OTHER";
+
+export type AcceptOccupation = "STUDENT" | "PROFESSIONAL" | "ANY";
