+++ backend/src/pages/api/identityDocs/[uid]/WORK_ID/readImage/index.ts
@@ -11,7 +11,7 @@ import HeaderTypes from "@/types/HeaderTypes";
 /**
  * ```
  * request = "GET /api/identityDocs/[uid]/WORK_ID/readImage?size=(small|medium|large)&b64=boolean"
- * response = 301 to "Content-Type: image/(jpeg|png)"
+ * response = text/plain (base64) is b64 = true, else image/*
  * ```
  */
 export default WithMiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
