+++ backend/src/pages/api/profile/[uid]/readImage/index.ts
@@ -10,7 +10,7 @@ import HeaderTypes  from "@/types/HeaderTypes";
 /**
  * ```
  * request = "GET /api/profile/[uid]/readImage?size=(small|medium|large)&b64=boolean"
- * response = 301 to "Content-Type: image/(jpeg|png)"
+ * response = text/plain (base64) is b64 = true, else image/*
  * ```
  */
 export default WithMiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
